Refactor ConfigurationAPI.ts
Refactor `src/services/backend/api/ConfigurationAPI.ts` into `ConfigGet.ts`, `ConfigSave.ts`, `ConfigValidate.ts`, and `ConfigReset.ts`, updating the original file to orchestrate these new components. All functionality preserved.
